Attempt to fix compilation in Guix.
