+++ src/client-api/bigint.cpp
@@ -1,0 +1,15 @@
+#include "univalue.h"
+
+// The client will parse objects created by this function as JS bigints.
+UniValue BigInt(std::string s) {
+    UniValue r = UniValue::VOBJ;
+    r.pushKV("bigint", s);
+    return r;
+}
+
+UniValue BigInt(uint64_t n) {
+    std::ostringstream s;
+    s << n;
+
+    return BigInt(s.str());
+}
